refactor(feature): split feature serialization codecs into per-family files
Keep the recipe dispatch lean and each file under the size limit as more feature
codecs land: extrude and the dress-up family move to their own files, leaving
serialize.go with the FeatureData shape, the marshal/restore dispatch, and the
shared key/scalar helpers. No behaviour change.
