+++ model/feature/serialize.go
@@ -3,7 +3,6 @@
 package feature
 
 import (
-	"encoding/base64"
 	"fmt"
 
 	"github.com/Oblikovati/oblikovati/kernel/ops"
@@ -40,39 +39,6 @@ type FeatureData struct {
 
 	BoundaryPatch *BoundaryPatchData `yaml:"boundaryPatch,omitempty"`
 	RuledSurface  *RuledSurfaceData  `yaml:"ruledSurface,omitempty"`
-}
-
-// EdgeDressData is an edge-based dress-up (fillet radius / chamfer distance): the
-// picked edges as reference keys plus the scalar value. The keys are base64 because
-// they are opaque lineage-derived bytes — the one non-text field (ADR-0020); they
-// re-bind to the regenerated edges after recompute (kernel topo FindEdgeByKey).
-type EdgeDressData struct {
-	Edges []string `yaml:"edges"`
-	Value float64  `yaml:"value"`
-}
-
-// FaceDressData is a face-based dress-up (shell thickness / draft angle): the picked
-// faces as reference keys plus the scalar value.
-type FaceDressData struct {
-	Faces []string `yaml:"faces"`
-	Value float64  `yaml:"value"`
-}
-
-// ThreadData tags a single cylindrical face (reference key) with a thread designation.
-type ThreadData struct {
-	Face        string `yaml:"face"`
-	Designation string `yaml:"designation"`
-}
-
-// ExtrudeData is an extrude's recipe: which sketch profile, the boolean operation, and
-// the distance extent. The distance is the evaluated growth (a fixed value on reopen;
-// parametric distance expressions arrive with the dimension-driven extent API).
-type ExtrudeData struct {
-	Sketch    int     `yaml:"sketch"`
-	Profile   int     `yaml:"profile"`
-	Operation string  `yaml:"operation"`
-	Distance  float64 `yaml:"distance"`
-	Taper     float64 `yaml:"taper,omitempty"`
 }
 
 // SketchIndexer maps between a sketch pointer and its index in the part, so a feature
@@ -183,27 +149,6 @@ func serializeFeature(pf *PartFeature, sk SketchIndexer, idx map[ID]int) (Featur
 	return fd, nil
 }
 
-func serializeExtrude(def *ExtrudeDefinition, sk SketchIndexer) (*ExtrudeData, error) {
-	if def.Extent.Type != DistanceExtent {
-		return nil, fmt.Errorf("only distance extents are serializable (got extent type %d)", def.Extent.Type)
-	}
-	idx, ok := sk.IndexOf(def.Sketch)
-	if !ok {
-		return nil, fmt.Errorf("extrude references a sketch that is not in the part")
-	}
-	op, err := operationName(def.Operation)
-	if err != nil {
-		return nil, err
-	}
-	return &ExtrudeData{
-		Sketch:    idx,
-		Profile:   def.ProfileIndex,
-		Operation: op,
-		Distance:  def.Extent.distance(),
-		Taper:     def.Taper,
-	}, nil
-}
-
 // ApplyRecipe rebuilds the feature program from its serialized form, in order. It
 // tracks the features restored so far so a pattern/mirror can resolve the earlier
 // features it replicates (recorded as program indices). The caller recomputes
@@ -285,103 +230,6 @@ func buildFeature(fs *PartFeatures, fd FeatureData, sk SketchIndexer, restored [
 		return nil, fmt.Errorf("no restore codec for feature kind %q", fd.Kind)
 	}
 }
-
-// requireExtrude restores an extrude, erroring on a missing payload.
-func requireExtrude(fs *PartFeatures, ed *ExtrudeData, sk SketchIndexer) (*PartFeature, error) {
-	if ed == nil {
-		return nil, fmt.Errorf("extrude feature is missing its payload")
-	}
-	return restoreExtrude(fs, ed, sk)
-}
-
-// dressInputs is the decoded (keys, value) pair shared by edge/face dress-ups.
-type dressInputs struct {
-	keys  [][]byte
-	value float64
-}
-
-func requireEdgeDress(d *EdgeDressData, kind string) (dressInputs, error) {
-	if d == nil {
-		return dressInputs{}, fmt.Errorf("%s feature is missing its payload", kind)
-	}
-	keys, err := decodeKeys(d.Edges)
-	if err != nil {
-		return dressInputs{}, err
-	}
-	return dressInputs{keys: keys, value: d.Value}, nil
-}
-
-func requireFaceDress(d *FaceDressData, kind string) (dressInputs, error) {
-	if d == nil {
-		return dressInputs{}, fmt.Errorf("%s feature is missing its payload", kind)
-	}
-	keys, err := decodeKeys(d.Faces)
-	if err != nil {
-		return dressInputs{}, err
-	}
-	return dressInputs{keys: keys, value: d.Value}, nil
-}
-
-func restoreExtrude(fs *PartFeatures, ed *ExtrudeData, sk SketchIndexer) (*PartFeature, error) {
-	skt, ok := sk.At(ed.Sketch)
-	if !ok {
-		return nil, fmt.Errorf("extrude references sketch index %d, which does not exist", ed.Sketch)
-	}
-	op, err := parseOperation(ed.Operation)
-	if err != nil {
-		return nil, err
-	}
-	dist := ed.Distance
-	pf := NewExtrudeFeatures(fs).AddByDistanceExtent(skt, ed.Profile, op, func() float64 { return dist })
-	if ed.Taper != 0 {
-		pf.feature.(*ExtrudeFeature).def.Taper = ed.Taper
-	}
-	return pf, nil
-}
-
-// encodeKeys / decodeKeys base64-encode reference keys (opaque lineage bytes) so they
-// stay valid text in the YAML document (ADR-0020); they re-bind after recompute.
-func encodeKeys(keys [][]byte) []string {
-	out := make([]string, len(keys))
-	for i, k := range keys {
-		out[i] = encodeKey(k)
-	}
-	return out
-}
-
-func decodeKeys(encoded []string) ([][]byte, error) {
-	out := make([][]byte, len(encoded))
-	for i, e := range encoded {
-		k, err := decodeKey(e)
-		if err != nil {
-			return nil, err
-		}
-		out[i] = k
-	}
-	return out, nil
-}
-
-func encodeKey(k []byte) string { return base64.StdEncoding.EncodeToString(k) }
-
-func decodeKey(s string) ([]byte, error) {
-	k, err := base64.StdEncoding.DecodeString(s)
-	if err != nil {
-		return nil, fmt.Errorf("reference key %q is not valid base64: %w", s, err)
-	}
-	return k, nil
-}
-
-// evalFloat reads a dress-up's scalar (a closure, typically a parameter); a nil closure
-// reads as 0. constFloat is its inverse for restore — a fixed value (parametric
-// dress-up values arrive with the dimension-driven API, like extrude distance).
-func evalFloat(fn func() float64) float64 {
-	if fn == nil {
-		return 0
-	}
-	return fn()
-}
-
-func constFloat(v float64) func() float64 { return func() float64 { return v } }
 
 // evalInt / constInt are the integer counterparts for pattern counts.
 func evalInt(fn func() int) int {
